+++ refresh/templates/services/alertnotification/importDependency.swift
@@ -1,0 +1,1 @@
+        .Package(url: "https://github.com/IBM-Swift/alert-notification-sdk.git", majorVersion: 1),