+++ refresh/templates/services/alertnotification/initializeBluemixService.swift
@@ -1,0 +1,3 @@
+    let alertNotificationService = try manager.getAlertNotificationService(name: "<%- service.name %>")
+
+    serviceCredentials = ServiceCredentials(service: alertNotificationService)