+++ refresh/templates/services/watsonconversation/initializeBluemixService.swift
@@ -1,0 +1,3 @@
+    let service = try manager.getWatsonConversationService(name: "<%- service.name %>")
+
+    conversation = Conversation(service: service, version: "<%- service.version %>")